refactor(navigation): rename TopNav component to TopNavigation
Renamed TopNav.svelte to TopNavigation.svelte for improved clarity and
consistency. Updated all imports, exports, and references across:
- Component imports in +layout.svelte
- Export declarations in navigation index.ts
- Documentation files and architecture diagrams

No functionality changes, purely a naming refactor for better semantic clarity.
