+++ frontend/src/lib/components/navigation/TopNavigation.svelte
