+++ frontend/src/lib/components/navigation/index.ts
@@ -1,3 +1,3 @@
 export { default as LinkList } from './LinkList.svelte';
 export { default as ChatTrigger } from './ChatTrigger.svelte';
-export { default as TopNav } from './TopNav.svelte';
+export { default as TopNavigation } from './TopNavigation.svelte';
